Add item requests

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -9,6 +9,7 @@ import ru.practicum.shareit.user.repository.UserRepository;
 import ru.practicum.shareit.user.userDto.UserDto;
 import ru.practicum.shareit.user.userDto.UserMapper;
 
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -18,37 +19,44 @@ public class UserServiceImpl implements UserService {
     private UserRepository userRepository;
 
     @Override
+    @Transactional
     public UserDto create(UserDto userDto) {
-        return UserMapper.touserDto(userRepository.save(UserMapper.toUser(userDto)));
+        return UserMapper.toUserDto(userRepository.save(UserMapper.toUser(userDto)));
     }
 
     @Override
+    @Transactional
     public UserDto patch(long id, UserDto userDto) {
         UserDto userForUpdate = getUserById(id);
         if (userDto.getEmail() != null && !userDto.getEmail().isBlank()) userForUpdate.setEmail(userDto.getEmail());
         if (userDto.getName() != null && !userDto.getName().isBlank()) userForUpdate.setName(userDto.getName());
-        return UserMapper.touserDto(userRepository.save(UserMapper.toUser(userForUpdate)));
+        return UserMapper.toUserDto(userRepository.save(UserMapper.toUser(userForUpdate)));
     }
 
     @SneakyThrows
     @Override
+    @Transactional
     public UserDto getUserById(long id) {
         User currentUser = userRepository.getReferenceById(id);
         if (currentUser == null) {
-            throw new DataNotFound("Пользователь c ID: " + id + " не найден!");
+            throw new DataNotFound("Пользователь не найден");
         }
-        return UserMapper.touserDto(currentUser);
+        return UserMapper.toUserDto(currentUser);
     }
 
     @Override
     public List<UserDto> getAllUsers() {
         return userRepository.findAll().stream()
-                .map(UserMapper::touserDto)
+                .map(UserMapper::toUserDto)
                 .collect(Collectors.toList());
     }
 
     @Override
     public void removeUser(long id) {
+        User currentUser = userRepository.getReferenceById(id);
+        if (currentUser == null) {
+            throw new DataNotFound("Пользователь c ID: " + id + " не найден!");
+        }
         userRepository.deleteById(id);
     }
 }
